requirements.txt
# ============================================
# Wise Code Watchers - Python Dependencies
# ============================================

# Web Framework
Flask>=3.0.0

# GitHub Integration
PyGithub>=2.1.1
PyJWT>=2.8.0
requests>=2.31.0

# LLM / LangChain Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# Environment Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# HTTP Client
httpx>=0.24.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Code Analysis & Parsing
ast>=3.9.0

# Utilities
python-dateutil>=2.8.0

# Optional: Advanced static analysis (Semgrep - install separately)
# semgrep>=1.45.0

# Optional: Security scanning tools
# bandit>=1.7.5
# safety>=2.3.0

# Development Dependencies (uncomment for development)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.5.0
